# apps/api/src/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "walk / dog API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/walks/active": {
      "get": {
        "tags": ["walks"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Active Walk" },
          "204": { "description": "No active walk" },
          "401": { "description": "Unauthenticated" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/v1/walks": {
      "post": {
        "tags": ["walks"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "201": { "description": "Created Walk" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthenticated" },
          "404": { "description": "Not found" },
          "409": { "description": "Conflict" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/v1/walks/{walkId}": {
      "get": {
        "tags": ["walks"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "walkId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Completed Walk Detail" },
          "401": { "description": "Unauthenticated" },
          "404": { "description": "Not found" },
          "500": { "description": "Internal server error" }
        }
      },
      "delete": {
        "tags": ["walks"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "walkId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Walk discarded" },
          "401": { "description": "Unauthenticated" },
          "404": { "description": "Not found" },
          "409": { "description": "Conflict" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/v1/walks/{walkId}/finish": {
      "post": {
        "tags": ["walks"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "walkId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 256 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Completed Walk" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthenticated" },
          "404": { "description": "Not found" },
          "409": { "description": "Conflict" },
          "500": { "description": "Internal server error" },
          "503": { "description": "Service unavailable" }
        }
      }
    },
    "/v1/walks/{walkId}/track-points": {
      "post": {
        "tags": ["walks"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "walkId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["recordedAt", "latitude", "longitude"],
                "properties": {
                  "recordedAt": { "type": "string", "format": "date-time" },
                  "latitude": { "type": "number" },
                  "longitude": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Accepted TrackPoint" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthenticated" },
          "404": { "description": "Not found" },
          "409": { "description": "Conflict" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/v1/walks/{walkId}/events": {
      "post": {
        "tags": ["walks"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "walkId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "eventId",
                  "participantDogId",
                  "type",
                  "occurredAt",
                  "latitude",
                  "longitude"
                ],
                "properties": {
                  "eventId": { "type": "string", "format": "uuid" },
                  "participantDogId": { "type": "string", "format": "uuid" },
                  "type": {
                    "type": "string",
                    "enum": ["pee", "poop", "sniff", "greet"]
                  },
                  "occurredAt": { "type": "string", "format": "date-time" },
                  "latitude": { "type": "number" },
                  "longitude": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Replayed Event" },
          "201": { "description": "Recorded Event" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthenticated" },
          "404": { "description": "Not found" },
          "409": { "description": "Conflict" },
          "500": { "description": "Internal server error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}